requirements.txt
tqdm
numpy
matplotlib
Pillow
tensorboardX
pandas
PyYAML
sklearn
torchaudio==0.6.0